#23: rename package
Following recommendations from https://blog.golang.org/package-names
